add coco package and add the yolo data generator

+++ utils.py
@@ -52,10 +52,10 @@ def convert_yolo_to_abs(
 
     grid_x, grid_y, x, y, w, h = yolo_loc
     return [
-        int(grid_cell_width * (grid_x + x)) - int(w * img_width) / 2,
-        int(grid_cell_height * (grid_y + y)) - int(h * img_height) / 2,
-        int(grid_cell_width * (grid_x + x)) + int(w * img_width) / 2,
-        int(grid_cell_height * (grid_y + y)) + int(h * img_height) / 2
+        int(int(grid_cell_width * (grid_x + x)) - int(w * img_width) / 2),
+        int(int(grid_cell_height * (grid_y + y)) - int(h * img_height) / 2),
+        int(int(grid_cell_width * (grid_x + x)) + int(w * img_width) / 2),
+        int(int(grid_cell_height * (grid_y + y)) + int(h * img_height) / 2)
     ]
